+++ CHANGELOG.md
@@ -1,3 +1,36 @@
+#### 0.5.0 (2021-03-24)
+
+##### Chores
+
+* **deps:**  force latest version ([a7d880ce](https://github.com/CodeTanzania/majifix-int-dawasco/commit/a7d880ce3f0707d51ca21084b4f4c7e27e9778c1))
+*  update examples ([aeea67bc](https://github.com/CodeTanzania/majifix-int-dawasco/commit/aeea67bc77d16a5becabf86cd33ccc66fcfc8a78))
+*  ignore husky config ([2a57f3cd](https://github.com/CodeTanzania/majifix-int-dawasco/commit/2a57f3cd1a5015146ecb1398f7eab20c25dc5608))
+*  add husky prepare task ([9aaca3f1](https://github.com/CodeTanzania/majifix-int-dawasco/commit/9aaca3f1026f69f1a674a45315b22323e0686965))
+*  improve package documentation ([9aadbf39](https://github.com/CodeTanzania/majifix-int-dawasco/commit/9aadbf39931648a07c60bef48969b3ac63918605))
+*  restrure to use es6 ([9de280a2](https://github.com/CodeTanzania/majifix-int-dawasco/commit/9de280a25bc47b134de842fda401a6c449c106a0))
+*  add common repository files ([c38298f8](https://github.com/CodeTanzania/majifix-int-dawasco/commit/c38298f8bbba4af120f8fe2859e2a33801c62087))
+*  add common config files ([15ebadd0](https://github.com/CodeTanzania/majifix-int-dawasco/commit/15ebadd0992bce1cf3add2e2a7ed22976eb84f08))
+
+##### New Features
+
+*  get pond bill number ([bd3a9052](https://github.com/CodeTanzania/majifix-int-dawasco/commit/bd3a905212b1d41086edea07ac193928eebc01ce))
+* fetch pond bill ([0cd7c331](https://github.com/CodeTanzania/majifix-int-dawasco/commit/0cd7c3313610384d056f00ae877151c8abe2c185))
+*  normalize api options ([2bc4d3d2](https://github.com/CodeTanzania/majifix-int-dawasco/commit/2bc4d3d2d49df1a698cd6c1c4d99ed8ba04141d5))
+*  add fetch account ([42a342dc](https://github.com/CodeTanzania/majifix-int-dawasco/commit/42a342dcaeec10e9b0c63c10627768df91156f18))
+*  get current and previous bills ([d552b766](https://github.com/CodeTanzania/majifix-int-dawasco/commit/d552b7663b3e64099f9bedc1bd6d3743883eda46))
+*  get customer details ([06f0eba8](https://github.com/CodeTanzania/majifix-int-dawasco/commit/06f0eba81383cac39522402a60c809d0614d489f))
+*  get account details ([2242346e](https://github.com/CodeTanzania/majifix-int-dawasco/commit/2242346e0d604d47ac1fae9904ed0ef5b4fdc580))
+*  add default constants ([8b3630e2](https://github.com/CodeTanzania/majifix-int-dawasco/commit/8b3630e222db09aec019ac9bbaf1a0ffc34dba51))
+
+##### Other Changes
+
+* //github.com/CodeTanzania/majifix-int-dawasco ([32dd8641](https://github.com/CodeTanzania/majifix-int-dawasco/commit/32dd86418c79257cb774f302f5b6d3285dadabd2))
+* **release:**  compile latest source && release notes" ([965e6b88](https://github.com/CodeTanzania/majifix-int-dawasco/commit/965e6b88909d46d425f6d8ba7212a4bc3ae96978))
+
+##### Tests
+
+*  add common bill fixtures ([4b5fbb62](https://github.com/CodeTanzania/majifix-int-dawasco/commit/4b5fbb625879e9298662a43654ab2f6ed02f0bc9))
+
 #### 0.4.0 (2019-08-20)
 
 ##### Chores